+++ src/app/(payment)/payment/_components/Checkout/CheckoutForm/parts/CheckoutAddress.module.scss
@@ -1,0 +1,34 @@
+@mixin default-layout {
+  @include flex-column(1.2rem);
+  @include pretendard-20-500;
+
+  padding: 2rem;
+  border-radius: 0.4rem;
+  background: $white;
+}
+
+.address-section {
+  @include default-layout;
+
+  .address-box {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    .address-wrap {
+      @include pretendard-16-400;
+
+      display: flex;
+      gap: 1.4rem;
+
+      .address-key {
+        color: $gray-40;
+      }
+
+      .address-key,
+      .address-value {
+        @include flex-column(0.4rem);
+      }
+    }
+  }
+}
